+++ reptor/api/TemplatesAPI.py
@@ -13,14 +13,6 @@ class TemplatesAPI(APIClient):
             self.reptor.get_config().get_server(), f"api/v1/findingtemplates/"
         )
 
-    def get_template_overview(self) -> typing.List[FindingTemplate]:
-        """Gets list of Templates"""
-        response = self.get(self.base_endpoint)
-        return_data = list()
-        for item in response.json()["results"]:
-            return_data.append(FindingTemplate(item))
-        return return_data
-
     def get_template(self, template_id: str) -> FindingTemplate:
         """Gets a single Template by ID"""
         response = self.get(urljoin(self.base_endpoint, template_id))
@@ -31,10 +23,10 @@ class TemplatesAPI(APIClient):
         url = urljoin(self.base_endpoint, f"{template_id}/export/")
         return self.post(url).content
 
-    def search(self, search_term) -> typing.List[FindingTemplate]:
+    def search(self, search_term: str = "") -> typing.List[FindingTemplate]:
         """Searches through the templates"""
-        response = self.get(urljoin(self.base_endpoint, f"?search={search_term}"))
-        return [FindingTemplate(item) for item in response.json()["results"]]
+        templates_raw = self.get_paginated(self.base_endpoint, params={"search": search_term})
+        return [FindingTemplate(template_raw) for template_raw in templates_raw]
 
     def get_templates_by_tag(self, tag: str) -> typing.List[FindingTemplate]:
         matched_templates = list()